ADDED : Watch Service to Watch Particular Service

+++ src/main/java/com/bridgelabz/FileWatchService.java
@@ -9,41 +9,37 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.stream.IntStream;
 
-public class NIOFileAPITest {
-    public static final String HOME = System.getProperty("user.home");
-    public static final String PLAY_WITH_NIO = "TempPlayground";
+public class NioFileApiTest {
+
+    public static String HOME = System.getProperty("user.home");
+    private static String PLAY_WITH_NIO="tempPlayGround";
 
     @Test
-    public void givenPath_whenChecked_thenConfirm() throws IOException {
-        //check file Exists
+    public void givenPathWhenRight_ShouldPass() throws IOException {
         Path homePath = Paths.get(HOME);
         Assert.assertTrue(Files.exists(homePath));
-
-        //Delete file and check file not exist
-        Path playPath = Paths.get(HOME + "/" + PLAY_WITH_NIO);
-        if (Files.exists(playPath))
-            FileUtils.deleteFiles(playPath.toFile());
+        Path playPath = Paths.get(HOME+"/"+PLAY_WITH_NIO);
+        if(Files.exists(playPath))FileUtils.deleteFiles(playPath.toFile());
         Assert.assertTrue(Files.notExists(playPath));
-
-        //Create Directory
-        Files.createDirectory(playPath);
+        Files.createDirectories(playPath);
         Assert.assertTrue(Files.exists(playPath));
 
-        //create file
-        IntStream.range(1, 10).forEach(counter -> {
-            Path tempFile = Paths.get(playPath + "/temp" + counter);
+        IntStream.range(1,10).forEach(fileNumber ->{
+            Path tempFile= Paths.get(playPath+"/temp"+fileNumber);
             Assert.assertTrue(Files.notExists(tempFile));
             try {
                 Files.createFile(tempFile);
-            } catch (IOException e) {
-                e.printStackTrace();
+
             }
-            Assert.assertTrue(Files.exists(tempFile));
+            catch (IOException e){
+                Assert.assertTrue(Files.exists(tempFile));
+            }
         });
 
         Files.list(playPath).filter(Files::isRegularFile).forEach(System.out::println);
         Files.newDirectoryStream(playPath).forEach(System.out::println);
-        Files.newDirectoryStream(playPath, path -> path.toFile().isFile() && path.toString().startsWith("temp")).forEach(System.out::println);
+        Files.newDirectoryStream(playPath,path -> path.toFile().isFile()&&path.toString().startsWith("temp")).forEach(System.out::println);
+
     }
 
     @Test
